create new file for reducers/actions testing
